feat: organize directory structure and exports
