+++ docs/src/content/config.ts
@@ -1,4 +1,4 @@
-import { docsSchema } from '@miracle-box/flint/schema';
+import { docsSchema } from '@miracle-box/flint/schemas';
 import { defineCollection } from 'astro:content';
 
 const docs = defineCollection({
