+++ docs/src/middleware.ts
@@ -1,4 +1,4 @@
-import { flintMiddlewares } from '@miracle-box/flint/middleware';
+import { flintMiddlewares } from '@miracle-box/flint/middlewares';
 import { sequence } from 'astro/middleware';
 
 export const onRequest = sequence(...flintMiddlewares);
