+++ docs/src/pages/[locale]/index.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from '@miracle-box/flint/layouts/base.astro';
+import { BlankPage as Layout } from '@miracle-box/flint/components';
 import Hero from '../../components/hero.astro';
 
 export { staticPathForEachLocale as getStaticPaths } from '@miracle-box/flint/utils';
