+++ packages/flint/404.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from './layouts/blank.astro';
+import Layout from './content-components/blank-page.astro';
 import { locales } from './utils/i18n-path';
 import { getPathname } from './utils/route';
