+++ packages/flint/components.ts
@@ -1,4 +1,5 @@
-export { default as Callout } from './components/content/callout.astro';
-export { default as IntroCard } from './components/content/intro-card.astro';
-export { default as IntroHeader } from './components/content/intro-header.astro';
-export { default as IntroPane } from './components/content/intro-pane.astro';
+export { default as BlankPage } from './content-components/blank-page.astro';
+export { default as Callout } from './content-components/callout.astro';
+export { default as IntroHeader } from './content-components/intro-header.astro';
+export { default as IntroPane } from './content-components/intro-pane.astro';
+export { default as IntroCard } from './content-components/intro-card.astro';
