+++ packages/flint/components/breadcrumb.astro
@@ -1,7 +1,7 @@
 ---
 import Config from 'virtual:flint/config';
-import { fetchDocs } from '../../utils/docs';
-import { getPagePath } from '../../utils/route';
+import { fetchDocs } from '../utils/docs';
+import { getPagePath } from '../utils/route';
 
 type Props = {
 	filePath: string;
