+++ packages/flint/components/footer.astro
@@ -1,6 +1,6 @@
 ---
 import Config from 'virtual:flint/config';
-import { getPathname } from '../../utils/route';
+import { getPathname } from '../utils/route';
 ---
 
 <footer class="footer">
