+++ packages/flint/components/navbar.astro
@@ -1,8 +1,8 @@
 ---
 import type { LocaleItem } from './locale-selector';
 import Config from 'virtual:flint/config';
-import { changePathLocale } from '../../utils/i18n-path';
-import { getPagePath } from '../../utils/route';
+import { changePathLocale } from '../utils/i18n-path';
+import { getPagePath } from '../utils/route';
 import { LocaleSelector } from './locale-selector';
 
 const { url } = Astro;
