+++ packages/flint/components/outdated-page-tip.astro
@@ -1,5 +1,5 @@
 ---
-import Callout from '../content/callout.astro';
+import Callout from '../content-components/callout.astro';
 
 const t = Astro.locals.translations;
 ---
