+++ packages/flint/components/outdated-translation-tip.astro
@@ -1,6 +1,6 @@
 ---
-import { changePathLocale } from '../../utils/i18n-path';
-import Callout from '../content/callout.astro';
+import Callout from '../content-components/callout.astro';
+import { changePathLocale } from '../utils/i18n-path';
 
 type Props = {
 	originalLocale: string;
