+++ packages/flint/components/page-frame.astro
@@ -1,9 +1,9 @@
 ---
 import 'modern-normalize/modern-normalize.css';
 import Config from 'virtual:flint/config';
-import Footer from '../components/layout/footer.astro';
-import Navbar from '../components/layout/navbar.astro';
 import '../styles/variables.css';
+import Footer from './footer.astro';
+import Navbar from './navbar.astro';
 
 const { locale } = Astro.locals;
 const t = Astro.locals.translations;
