+++ packages/flint/content-components/blank-page.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from './base.astro';
+import Layout from '../components/page-frame.astro';
 ---
 
 <Layout>
