+++ packages/flint/content-components/callout.astro
@@ -1,5 +1,5 @@
 ---
-import type { Icon } from '../../types';
+import type { Icon } from '../types';
 import WarningIcon from '~icons/octicon/alert-16';
 import TipIcon from '~icons/octicon/info-16';
 import SuccessIcon from '~icons/octicon/issue-closed-16';
