+++ packages/flint/content-components/intro-pane.astro
