--- packages/flint/i18n-utils.ts
@@ -1,2 +1,0 @@
-export type { FlintUserDictionary } from './schemas/i18n';
-export { t, tSchema } from './utils/i18n-translation';
