+++ packages/flint/i18n/dictionary.ts
@@ -1,0 +1,5 @@
+import type { FlintTranslation } from './schema';
+import en from './en';
+import zh from './zh';
+
+export const flintDict: Record<string, FlintTranslation> = { en, zh };
