+++ packages/flint/i18n/index.ts
@@ -1,5 +1,3 @@
-import type { FlintTranslation } from '../schemas/i18n';
-import en from './en';
-import zh from './zh';
-
-export const flintDict: Record<string, FlintTranslation> = { en, zh };
+export * from './schema';
+export * from './translation';
+export * from './dictionary';
