+++ packages/flint/i18n/translation.ts
@@ -1,9 +1,4 @@
-import { z } from 'astro/zod';
 import Config from 'virtual:flint/config';
-
-export function tSchema() {
-	return z.function().returns(z.string());
-}
 
 export function t(text: TemplateStringsArray) {
 	const tFunc = () => text.join('');
