+++ packages/flint/i18n/zh/index.ts
@@ -1,5 +1,5 @@
-import type { FlintTranslation } from '../../schemas/i18n';
-import { t, localeLabel } from '../../utils/i18n-translation';
+import type { FlintTranslation } from '../schema';
+import { t, localeLabel } from '../translation';
 
 const zh: FlintTranslation = {
 	site: {
