--- packages/flint/middleware.ts
@@ -1,3 +1,0 @@
-import { flintI18nMiddleware } from './middlewares/i18n';
-
-export const flintMiddlewares = [flintI18nMiddleware];
