+++ packages/flint/middlewares/i18n/index.ts
@@ -1,5 +1,6 @@
 import { defineMiddleware } from 'astro/middleware';
-import { CustomDict, FlintDict, getLocaleFromUrl } from '../utils';
+import { getLocaleFromUrl } from '../../utils/i18n-path';
+import { FlintDict, CustomDict } from './dict-builder';
 
 const flintDict = new FlintDict();
 const customDict = new CustomDict();
@@ -11,6 +12,5 @@ export const flintI18nMiddleware = defineMiddleware(async (ctx, next) => {
 	// @ts-expect-error We do not know about the user defined type
 	ctx.locals.customTranslations = customDict.useTranslation(locale);
 
-	const response = await next();
-	return response;
+	return next();
 });
