+++ packages/flint/middlewares/index.ts
@@ -1,0 +1,3 @@
+import { flintI18nMiddleware } from './i18n';
+
+export const flintMiddlewares = [flintI18nMiddleware];
