+++ packages/flint/page.astro
@@ -1,7 +1,7 @@
 ---
 import type { DocsPageProps as Props } from './types';
 import Config from 'virtual:flint/config';
-import Layout from './layouts/docs-content.astro';
+import Layout from './components/content-page.astro';
 import { fetchDocs, getFallbackBySlugs, getLocalDocs } from './utils/docs';
 import { locales } from './utils/i18n-path';
