--- packages/flint/schema.ts
@@ -1,1 +1,0 @@
-export { docsSchema } from './schemas/docs';
