+++ packages/flint/utils.ts
@@ -1,3 +1,3 @@
-export * from './utils/route';
-export * from './utils/i18n-dict';
+export { FlintConfigSchema, type RawFlintConfig } from './utils/config';
 export * from './utils/i18n-path';
+export * from './utils/route';
