+++ packages/flint/virtual.d.ts
@@ -8,5 +8,5 @@ declare module 'virtual:flint/config' {
 declare module 'virtual:flint/user-translation' {
 	export const customSchema: import('astro/zod').ZodUnknown;
 	export const customDict: Record<string, unknown>;
-	export const flintUserDict: Record<string, import('./i18n-utils').FlintUserDictionary>;
+	export const flintUserDict: Record<string, import('./i18n').FlintUserDictionary>;
 }
